Adding options for default_dir, resource_root, and root_dir
I'm working on an app which has a slightly different directory structure
than the one expected by this library. In particular, I'd like the
resources directory to be off the root of the project, and the
preferences directory to be in `~/.config/<my-addin>`.

This PR doesn't make those changes, which would be subjective and
represent a breaking change. Instead I've made default_dir configurable
via kwargs to the app constructor. I've also added a new option for
`resource_root`, which defaults to _command/resources_ to avoid any
breaking changes for existing add ins.